Finished JWT logic

+++ backend/services/authorService.js
@@ -22,14 +22,22 @@ const createAuthor = async body => {
     return await Author.create(body);
 }
 
-const updateAuthor = async (id, newPassword) => {
+const updateAuthor = async (id, newPassword, token) => {
+    const author = await Author.findById(id);
+    if (author._id?.toString() !== token.id)
+       return; 
+
     const salt = await bcrypt.genSalt(10);
     const hashedPassword = await bcrypt.hash(newPassword, salt);
 
     await Author.findByIdAndUpdate(id, { password: hashedPassword });
 }
 
-const deleteAuthor = async id => {
+const deleteAuthor = async (id, token) => {
+    const author = await Author.findById(id);
+    if (author._id?.toString() !== token.id)
+       return; 
+
     await Author.findByIdAndDelete(id);
 }
